+++ README.rst
@@ -1,12 +1,12 @@
-Cloud Storage Cache
-===================
+Storage Cache
+=============
 
 Installation
 ------------
 
 .. code-block:: bash
 
-    pip install cloud-storage-cache
+    pip install storage-cache
 
 
 How to use
@@ -14,9 +14,9 @@ How to use
 
 .. code-block:: python
 
-    from cloud_storage_cache import CloudStorageCacheFactory
-    from cloud_storage_cache.types import REDIS_STORAGE_CACHE
+    from storage_cache import StorageCacheFactory
+    from storage_cache.types import REDIS_STORAGE_CACHE
 
-    CloudStorageCache = CloudStorageCacheFactory.factory(REDIS_STORAGE_CACHE)
-    cloud_storage_cache = CloudStorageCache(
+    StorageCache = StorageCacheFactory.factory(REDIS_STORAGE_CACHE)
+    storage_cache = StorageCache(
         <redis_host>, <redis_port>, <redis_password>, <redis_db>, <key_format>)
