--- cloud_storage_cache/__init__.py
@@ -1,1 +1,0 @@
-from cloud_storage_cache.main import CloudStorageCacheFactory
