+++ storage_cache/__init__.py
@@ -1,0 +1,5 @@
+from storage_cache.main import StorageCacheFactory
+
+__all__ = (
+    "StorageCacheFactory",
+)
